rsvp fully works

+++ app/controllers/likes_controller.rb
@@ -1,13 +1,15 @@
 class LikesController < ApplicationController
-  # before_action :authenticate_user!
+  before_action :authenticate_user!
 
   def create
     event = Event.find params[:event_id]
     like = Like.new(user: current_user, event: event)
-    if like.save
-      redirect_to event, notice: 'Thanks for liking!'
+    if !can? :like, event
+      head :unauthorized
+    elsif like.save
+      redirect_to event, notice: 'You are going!'
     else
-      redirect_to event, alert: 'You already  liked the item.'
+      redirect_to event, alert: 'You are going already!'
     end
   end
 
@@ -15,9 +17,11 @@ class LikesController < ApplicationController
 
   def destroy
     like = Like.find params[:id]
-    like.destroy
-    redirect_to like.event, notice: 'Like removed!'
+    if can? :destroy, like
+      like.destroy
+      redirect_to like.event, notice: 'You are not going!'
+    else
+      head :unauthorized
+    end
   end
-
-
 end
